M03: move Aggregate analyzer to hpvsim/analyzers.py
Pulls the per-genotype any-aggregator out of sim.py into its own
analyzers module — keeps sim.py focused on Sim construction and gives
future analyzers a natural home. sim.py imports it back for the
auto-add default; __init__.py re-exports for parity with the other
top-level building blocks (HPV, SexualNetwork, CrossImmunity, etc).
